Refactor isort sorting to work on Eiter iterator
There was a pattern in `OrderedImportBlock` that assumed `Import` and `ImportFrom`
statements will always be sorted independently.

This pattern made `force-sort-within-sections` awkward to implement
(sorting has to occur on `OrderedImportBlock`).

This commit moves some things around to get a single stream of `Either<Import, ImportFrom>`.
It also introduces a single `cmp` function that can sort the stream based on the given rules.

Some of the additions here are not ideal.
Specifically:

- `import_as_to_from_import`
- `import_from_star_to_from_import`
- `from_import_to_from_import`

There are some differing intermediate data structures used in the categorization stage.
These structures are just nested tuples.

Maybe it would make sense to have a single struct when collecting import data,
and the category be a method on that struct?

That way we could do away with:

- The conversion functions.
- A lot of the iteration work.
- The `Either` enum.

I think you could just match on category for the final sort function.